refactor(test): extract shared E2E helpers, trim CSS-only tests, document limitations
**Shared helpers extracted (3 new files):**
- helpers/theme.ts: clickThemeToggle — extracted from 4 test files
- helpers/portfolio.ts: openFilterDrawer — extracted from 2 test files
- helpers/regulatory-map.ts: clickSvgPath, waitForMapReady — extracted from 5 files
- Consolidated setupAnalyticsMocking imports in about-page + analytics
  (replaced local redefinitions with existing shared helper)

**CSS-only tests trimmed:**
- filter-drawer-layering.test.ts: deleted 14 pure CSS property inspection
  tests (z-index, position, overflow, transition, pointer-events, width,
  flex layout), kept 7 behavioral tests (open/close/toggle/clear)

**Integration test documentation:**
- diligence-wizard-navigation.test.ts: documented that WizardNavigationSimulator
  replicates inline script logic that cannot be extracted without major refactor
- toc-component.test.ts: documented that functions mirror inline Astro
  template expressions
