+++ tests/e2e/portfolio-filtering.test.ts
@@ -1,24 +1,5 @@
 import { test, expect } from '@playwright/test';
-
-/**
- * Open the filter drawer and wait for its slide-in transition to complete.
- */
-async function openFilterDrawer(page: import('@playwright/test').Page): Promise<void> {
-  // Use evaluate to bypass WebKit hit-testing
-  await page.evaluate(() => {
-    (document.querySelector('[data-testid="portfolio-filter-toggle"]') as HTMLElement)?.click();
-  });
-
-  const drawer = page.locator('[data-testid="portfolio-filter-drawer"]');
-  await expect(drawer).toBeVisible({ timeout: 5000 });
-
-  await page.waitForFunction(() => {
-    const el = document.querySelector('[data-testid="portfolio-filter-drawer"]');
-    if (!el || !el.classList.contains('open')) return false;
-    const right = parseFloat(window.getComputedStyle(el).right);
-    return right >= -1;
-  }, { timeout: 5000 });
-}
+import { openFilterDrawer } from './helpers/portfolio';
 
 test.describe('Portfolio Filtering - DOM Integration Tests', () => {
   test.beforeEach(async ({ page }) => {
