+++ tests/e2e/theme-toggle.test.ts
@@ -1,18 +1,5 @@
 import { test, expect } from '@playwright/test';
-
-/**
- * Click the theme toggle via dispatchEvent.
- * WebKit's hit-testing can fail on the toggle due to the footer's z-index: 0
- * stacking context + the large font-size creating an oversized bounding box.
- * dispatchEvent bypasses Playwright's coordinate-based click.
- */
-async function clickThemeToggle(page: import('@playwright/test').Page): Promise<void> {
-  await page.evaluate(() => {
-    document.getElementById('themeToggle')?.dispatchEvent(
-      new MouseEvent('click', { bubbles: true })
-    );
-  });
-}
+import { clickThemeToggle } from './helpers/theme';
 
 test.describe('Theme Toggle Journey', () => {
   test.beforeEach(async ({ page }) => {
